requirements.txt
pyyaml
schedule
requests
openai
dashscope
beautifulsoup4
feedparser
streamlit
pandas
psutil
selenium
playwright
